Setup typescript
