--- index.js
@@ -1,20 +1,0 @@
-const makeModelsController = require('./app/controllers/ModelsController');
-const setupController = require('./app/controllers/RadiksController');
-const { getDB } = require('./app/database/mongodb');
-
-const setup = async (config = {}) => {
-  const db = await getDB(config.mongoDBUrl);
-  const newConfig = { ...config };
-  if (!config.maxLimit) {
-    newConfig.maxLimit = 1000;
-  }
-  const controller = setupController(db, newConfig);
-  return controller;
-};
-
-module.exports = {
-  makeModelsController,
-  setup,
-  getDB,
-  setupController,
-};
